fix: move logo to elements-core
